Moved CodeQLTypes to ./Types
